Correct errors, add commit for every creation, save the objects in json files
